fix: Fixed Cross-site request forgery vulnerabilities on person update
Implemented the usage of a CSRF token to stop 3rd party access of user information updating.
The token is initialized on session creation and the value of the requestinng source's token is checked every time updatePerson() is called.
Addded the method of Cross-site request forgery defense into the documentation.
Fixed typo in the SQL injection documentation.